refactor kds, kds_wasm flows and tests+ci
