+++ src/kds.rs
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-#[cfg(not(any(feature = "online", target_arch = "wasm32")))]
-compile_error!("kds module requires either the 'online' feature or wasm32 target");
+#[cfg(not(feature = "kds"))]
+compile_error!("kds module requires the 'kds' feature");
 
 use crate::crypto::{Certificate, CertificateBackend, Crypto};
 use crate::snp;
