+++ tests/tests_native.rs
@@ -17,55 +17,53 @@ pub fn init_logger() {
     });
 }
 
-#[cfg(feature = "online")]
-mod online {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_verify_milan_attestation() {
-        init_logger();
-        common::verify_milan_attestation()
-            .await
-            .expect("Verification call failed");
-    }
-
-    #[tokio::test]
-    async fn test_verify_genoa_attestation() {
-        init_logger();
-        common::verify_genoa_attestation()
-            .await
-            .expect("Verification call failed");
-    }
-
-    #[tokio::test]
-    async fn test_verify_turin_attestation() {
-        init_logger();
-        common::verify_turin_attestation()
-            .await
-            .expect("Verification call failed");
-    }
-}
-
-/// Offline verification tests (sync, uses pinned ARKs)
-#[cfg(sync_crypto)]
-mod offline {
-    use super::*;
-
-    #[test]
-    fn test_suite() {
-        init_logger();
-        common::test_verify_attestation_suite();
-    }
-}
-
-/// Offline verification tests (async, uses pinned ARKs)
 #[cfg(async_crypto)]
-mod offline_async {
+mod r#async {
     use super::*;
 
     #[tokio::test]
     async fn test_suite() {
         init_logger();
         common::test_verify_attestation_suite_async().await;
+    }
+
+    #[cfg(feature = "kds")]
+    mod kds {
+        use super::*;
+
+        #[tokio::test]
+        async fn test_verify_milan_attestation() {
+            init_logger();
+            common::verify_milan_attestation()
+                .await
+                .expect("Verification call failed");
+        }
+
+        #[tokio::test]
+        async fn test_verify_genoa_attestation() {
+            init_logger();
+            common::verify_genoa_attestation()
+                .await
+                .expect("Verification call failed");
+        }
+
+        #[tokio::test]
+        async fn test_verify_turin_attestation() {
+            init_logger();
+            common::verify_turin_attestation()
+                .await
+                .expect("Verification call failed");
+        }
+    }
+}
+
+#[cfg(sync_crypto)]
+mod sync {
+    use super::*;
+
+    #[test]
+    fn test_suite() {
+        init_logger();
+        common::test_verify_attestation_suite();
     }
 }
